Add some basic scaffolding, ping and restart commands
This is just some basic scaffolding to set up the dependencies, and get
the main login/message handler out of the way. Right now the auth is
just "are you a user defined in an environment variable", and only the
most basic commands have been added. Ping is just to make sure we're
talking to discord at all, and restart is just a proof of concept of
working with the Heroku CLI for this.

The actual handling of the output was surprisingly tricky. We want
everything to be line buffered (so we don't hit Discord's length limit),
and we want bot stdout and stderr like you'd get from the terminal.
Node's streams are... not really meant for this sort of thing. It became
much simpler once I started working with JS async iterables, but there's
no way to just mix two iterables together so we have to write that
ourselves too. At least the result was somewhat easy to test.