Rewrite it in Rust
As requested in the last team meeting, this is the result of
investigating the serenity library for our purposes allowing us to write
this in Rust instead of JS (which for whatever reason didn't end up on
my radar when we did the initial spike). In terms of maintenance and
capabilities, the library certainly meets what we need.

This mostly behaves identically to the JS version, with the notable
exception that command output is only displayed once the command has
finished running, instead of streaming as lines come in. tokio-process
doesn't compile on the latest nightly. It seem prudent to at least wait
until 1.39 is in beta to switch over to async/await for this.

This is a naive port of the code. It does not use Serenity's built-in
command framework (which we may want to use, but it's very proc-macro
heavy).